moved PROPERTIES_LIST_VALUE_REGEX
