+++ wiggle_query_language/clauses/parsing_helpers/parse_statement_checks.py
@@ -11,12 +11,11 @@ from exceptions.wql.parsing import (
 from wiggle_query_language.clauses.regexes.make.make_patterns import (
     ILLEGAL_CHARS_REGEX,
     MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX_REGEX,
-    # CHECK_PARAMS_SYNTAX_REGEX,
-    PARAM_LIST_VALUE_REGEX,
 )
 
 from wiggle_query_language.clauses.regexes.patterns.properties import (
-    CHECK_PARAMS_SYNTAX_REGEX,
+    CHECK_PROPERTIES_SYNTAX_REGEX,
+    PROPERTIES_LIST_VALUE_REGEX,
 )
 from wiggle_query_language.clauses.regexes.patterns.relationships import (
     RELATIONSHIP_DIR_CHECK_REGEX,
@@ -46,7 +45,7 @@ def check_node_rel_properties(stmt_matches: list[str]) -> True:
     """
 
     for stmt in stmt_matches:
-        if not (param_string := CHECK_PARAMS_SYNTAX_REGEX.findall(stmt)):
+        if not (param_string := CHECK_PROPERTIES_SYNTAX_REGEX.findall(stmt)):
             continue
 
         for param_match_in in param_string:
@@ -57,11 +56,11 @@ def check_node_rel_properties(stmt_matches: list[str]) -> True:
             )
 
             # remove the list from the params
-            params_sans_list = PARAM_LIST_VALUE_REGEX.sub("", param_match)
+            params_sans_list = PROPERTIES_LIST_VALUE_REGEX.sub("", param_match)
             check_property_syntax(params_sans_list)
 
             # check lists
-            if params_lists := PARAM_LIST_VALUE_REGEX.findall(param_match):
+            if params_lists := PROPERTIES_LIST_VALUE_REGEX.findall(param_match):
                 for params_list in params_lists:
                     try:
                         eval(params_list)
